[Refactor] fix ruff issues.
