+++ avs/mic_thread.py
