+++ avs/windows/main_window.py
